requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-telegram-bot>=20.0
python-dotenv>=0.19.0
httpx>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
loguru>=0.5.3
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
alembic>=1.12.0